Switch to vertical search results in React app
Due to flexbox bugs in Chrome and Firefox, it's difficult to acheive the
desired horizontal layout of the search results. The bug is that when you
have a flex-direction:column inline-flexbox with multiple columns, the
width of the flexbox element is only the width of a single column.

See these bugs for more details:

Chrome:
https://code.google.com/p/chromium/issues/detail?id=247963

Firefox:
https://bugzilla.mozilla.org/show_bug.cgi?id=995020