Fix blank ListView bug in React app
The ListView displayed fine on initialization. However, when navigating
back to the page with the ListView, the ListView would be blank. The
problem is that there's a WinJS bug when trying to reuse a layout object
(the React app only creates 1 layout object and attempts to use it when
recreating the ListView upon navigation).

To workaround the WinJS bug, we utilize the { type: ... } contract for
layouts rather than creating one ourselves. This way each time we create a
new ListView, the ListView creates a new layout based on our description.

Reference to WinJS bug:
https://github.com/winjs/winjs/issues/1081